clippy: trim crate-level allow block + fix 14 mechanical lints
Drop 16 clippy/rustc allows from src/lib.rs that no longer fire — they
were originally added during the bring-up sweep but the underlying call
sites have since been refactored. Fix the 14 warnings that surfaced
when the allows were temporarily removed:

* needless_borrow x6 — `decode_celt_body` takes `rc: &mut RangeDecoder`,
  so passing `&mut rc` to inner helpers double-borrowed. Drop the extra
  `&mut` and the redundant `mut` on the parameter.
* manual_memcpy x3 — inner-loop `for i in 0..n { dst[i] = src[i] }`
  patterns swapped for `copy_from_slice`.
* unnecessary_cast x3 — drop `as i32` / `as u32` where the source is
  already that type.
* precedence x1 — parenthesise the `* 8 << BITRES` expression to make
  the `<<` precedence explicit.
* unused_mut x1 — drop `let mut` where the binding is never reassigned.

Remaining 4 allows (`dead_code`, `needless_range_loop`,
`excessive_precision`, `too_many_arguments`) still fire dozens of times
across the SILK/CELT pipelines and would each be a focused refactor.

Net: `cargo clippy --all-targets -- -D warnings` stays green; future
regressions in any of the 16 dropped lint categories now fail CI rather
than being silently masked. All 99 tests continue to pass.

+++ src/silk/synth.rs
@@ -136,7 +136,7 @@ pub fn upsample_to_48k(samples: &[f32], src_rate: u32) -> Vec<f32> {
         16_000 => upsample(samples, 3),
         24_000 => upsample(samples, 2),
         48_000 => samples.to_vec(),
-        _ => upsample(samples, 48_000 / src_rate as u32),
+        _ => upsample(samples, 48_000 / src_rate),
     }
 }
